+++ src/pages/Mypage/PopupPrivacyPolicy.tsx
@@ -1,5 +1,5 @@
 import { useTranslation } from "react-i18next";
-import Modal from "components/modal/Modal";
+import Modal from "components/Modal";
 import theme from "tools/theme";
 
 import { ReactComponent as TaxiLogo } from "static/assets/TaxiLogo.svg";
