+++ src/pages/Myroom/R2Myroom.jsx
@@ -5,8 +5,8 @@ import Title from "components/Title";
 import WhiteContainer from "components/WhiteContainer";
 import ChatHeaderBody from "pages/Chatting/Header/HeaderBody";
 import SideChat from "pages/Chatting/SideChat";
-import Room from "components/room/Room";
-import Pagination, { PAGE_MAX_ITEMS } from "components/pagination/Pagination";
+import Room from "components/Room";
+import Pagination, { PAGE_MAX_ITEMS } from "components/Pagination";
 import RLayout from "components/RLayout";
 import Empty from "components/Empty";
 import DottedLine from "components/DottedLine";
